Fix initial API deploys, communication with pg_bouncer instance.

+++ infrastructure/deploy.sh
@@ -218,21 +218,26 @@ chmod 600 data-refinery-key.pem
 API_IP_ADDRESS=$(terraform output -json api_server_1_ip | jq -c '.value' | tr -d '"')
 echo "Restarting API with latest image."
 
-ssh -o StrictHostKeyChecking=no \
-    -i data-refinery-key.pem \
-    ubuntu@$API_IP_ADDRESS  "docker pull $DOCKERHUB_REPO/$API_DOCKER_IMAGE"
+container_running=$(ssh -o StrictHostKeyChecking=no \
+                        -i data-refinery-key.pem \
+                        ubuntu@$API_IP_ADDRESS  "docker ps" | grep dr_api)
 
-ssh -o StrictHostKeyChecking=no \
-    -i data-refinery-key.pem \
-    ubuntu@$API_IP_ADDRESS "docker rm -f dr_api"
+if [[ ! -z $container_running ]]; then
+    ssh -o StrictHostKeyChecking=no \
+        -i data-refinery-key.pem \
+        ubuntu@$API_IP_ADDRESS  "docker pull $DOCKERHUB_REPO/$API_DOCKER_IMAGE"
 
-scp -o StrictHostKeyChecking=no \
-    -i data-refinery-key.pem \
-    api-configuration/environment ubuntu@$API_IP_ADDRESS:/home/ubuntu/environment
+    ssh -o StrictHostKeyChecking=no \
+        -i data-refinery-key.pem \
+        ubuntu@$API_IP_ADDRESS "docker rm -f dr_api"
 
-ssh -o StrictHostKeyChecking=no \
-    -i data-refinery-key.pem \
-    ubuntu@$API_IP_ADDRESS "docker run \
+    scp -o StrictHostKeyChecking=no \
+        -i data-refinery-key.pem \
+        api-configuration/environment ubuntu@$API_IP_ADDRESS:/home/ubuntu/environment
+
+    ssh -o StrictHostKeyChecking=no \
+        -i data-refinery-key.pem \
+        ubuntu@$API_IP_ADDRESS "docker run \
        --env-file environment \
        -e DATABASE_HOST=$DATABASE_HOST \
        -e DATABASE_NAME=$DATABASE_NAME \
@@ -246,6 +251,7 @@ ssh -o StrictHostKeyChecking=no \
        -p 8081:8081 \
        --name=dr_api \
        -it -d $DOCKERHUB_REPO/$API_DOCKER_IMAGE /bin/sh -c /home/user/collect_and_run_uwsgi.sh"
+fi
 
 
 echo "Deploy completed successfully."
